+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix a boilerplate issue when using options (extra comma)
 - fix handling of option `STRING` parameters
 - introduce delivery processors
+- introduce `VPriv::take()`
 
 # 0.0.6 (2021-12-23)